pyproject.toml
[tool.poetry]
name = "emendis"
version = "0.1.0"
description = ""
authors = ["Alexandre Kandalintsev <exe.sre@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10, <3.12"
fastapi = "^0.93.0"
uvicorn = {extras = ["standard"], version = "^0.20.0"}
orjson = "^3.8.7"
sqlalchemy = "^2.0.5.post1"
pyproject-flake8 = "^6.0.0.post1"
fastapi-filter = {extras = ["sqlalchemy"], version = "^0.5.3"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
fastapi-auth-middleware = "^1.0.2"


[tool.poetry.group.dev.dependencies]
poetry = "^1.4.0"
httpx = "^0.23.3"
pytest = "^7.2.2"
pdbpp = "^0.10.3"
black = "^23.1.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
alembic = "^1.10.2"
ipython = "^8.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
emendis = 'emendis.cli:cli'

[tool.flake8]
max-line-length = 120